Add files via upload
A server-side script that allows the player to purchase a particle effect for their character.  So long as the player can afford the price of the particles, subtract the given amount of currency from the player's game data, and add a copy of the ParticleEmitter object into the player's data as proof of purchase.  This script handles requests from the player to purchase, equip, and unequip any particle effect from the shop.